v2.0.0
Version 2.0.0 is out, here's whats new.

• Cook command (feature request by Rashaun) is now added.
• Phrases folder is to allow a user to add in items that they qant in the bot.
• Make it so that the owner is detected as botconfig.owner, not botconfig.prefix.
• Fix some role detection.

Not much bug fixing. Good luck finding bugs :3

+++ commands/fun/qt.js
@@ -26,6 +26,18 @@ var user = message.mentions.members.first() || message.guild.members.get(args[0]
 
 if(!user) return message.channel.send(`You need to call someone a qt, please \`ping\` someone or use their \`ID\`!`)
 else
+
+var nou = new Discord.RichEmbed()
+    .setTitle(`New QT Incoming ✨`)
+    .setAuthor(`${message.author.tag}`, message.author.avatarURL)
+    .addField(`QT Tag`, message.author)
+    .addField(`ID`, message.author.id)
+    .setColor(`#ff00bb`)
+    .setTimestamp()
+
+if(user.id == message.client.user.id) return message.channel.send(nou)
+else
+
 message.delete()
 var embed = new Discord.RichEmbed()
     .setTitle(`New QT Incoming ✨`)
