+++ commands/fun/remind.js
@@ -55,13 +55,13 @@ switch (tb) {
     case 'w':
         time = time * 1000 * 60 * 60 * 24 * 7;
         break
-    case 'y':
-        time = time * 1000 * 60 * 60 * 24 * 365;
-        break
 	default:
 		time = time * 1000;
 		break;
 }
+
+if(time > 2147483647) return message.channel.send(`This value is too high! Please go a little lower!`)
+else
 
 message.delete()
 message.channel.send(`📌 Ok, now reminding you in **${args[0]}** to do ${reason}. Please enable DM's in this server to recieve your reminder.`)
